Refactored code in to separated files and Add test case for who wirte the posts
