+++ spec/future/users_show_spec.rb
@@ -11,35 +11,6 @@ RSpec.feature 'Users', type: :feature do
     @last_post = Post.create(author: @user, title: 'Hello, four', text: 'This is my last post')
   end
 
-  describe 'user#index' do
-    before(:each) do
-      @user2 = User.create(name: 'Bill Gates',
-                           photo: 'https://somepic.potraits/0012',
-                           bio: 'Billonaire')
-      @user2_first_post = Post.create(author: @user2, title: 'Hello, there', text: 'This is my first post')
-      visit users_path
-    end
-    it 'should show username for each user' do
-      expect(page).to have_content('David')
-      expect(page).to have_content('Bill Gates')
-    end
-
-    it "should show user's profile picture" do
-      expect(page.body).to include('https://somepic.potraits/5158')
-      expect(page.body).to include('https://somepic.potraits/0012')
-    end
-
-    it 'should show the number of posts for each user' do
-      expect(page).to have_content('Number of posts: 4')
-      expect(page).to have_content('Number of posts: 1')
-    end
-
-    it 'should redirect the page to a user show page' do
-      click_link(@user2.name)
-      expect(current_path).to eq(user_path(@user2))
-    end
-  end
-
   describe 'user#show' do
     before(:each) do
       visit user_path(@user)
@@ -75,7 +46,7 @@ RSpec.feature 'Users', type: :feature do
       expect(page).to have_content('Pagination')
     end
 
-    it 'should redirect to the user posts page' do
+    it 'should redirect to the user posts page when the post title is clicked' do
       click_link(@first_post.title)
       expect(current_path).to eq(user_post_path(@user, @first_post))
     end
